Raised max map size to 192
git-svn-id: svn://localhost/trunk@2321 a386ff7f-d321-c144-bb17-e251d28fee14

+++ KM_FogOfWar.pas
@@ -3,13 +3,14 @@ unit KM_FogOfWar;
 interface
 uses Classes, Math, KM_Defaults, KM_CommonTypes, KM_Points;
 
+
 {These are mission specific settings and stats for each player}
 type
   TKMFogOfWar = class
   private
     fAnimStep:cardinal;
-    MapX:byte;
-    MapY:byte;
+    MapX: integer;
+    MapY: integer;
     Revelation:array of array of record
       //Lies within range 0, TERRAIN_FOG_OF_WAR_MIN..TERRAIN_FOG_OF_WAR_MAX.
       Visibility:byte;
